Software/tkinter-usb-app/requirements.txt
pyserial
matplotlib
numpy
isort